requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
greenlet==3.2.2
grpcio==1.71.0
# grpcio-status==1.65.1 # Unpin to allow resolution
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
huggingface-hub==0.24.1
idna==3.10
jiter==0.9.0
mem0ai==0.1.99
monotonic==1.6
numpy==2.2.5
openai==1.78.1
portalocker==2.10.1
posthog==3.25.0
# protobuf>=5.26.1,<6.0 # Unpin to allow resolution
protobuf # Let pip choose
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==310; sys_platform == 'win32'
qdrant-client==1.14.2
# requests==2.32.3 # Unpin to allow resolution
requests
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
python-dotenv
# supabase-py # Incorrect package name
# supabase-client==2.5.2 # Pin to a modern version
supabase # Correct PyPI package name for the official client
pgvector
asyncpg
aiohttp # Re-added, will be pre-installed in Dockerfile
# aiohttp # Added for supabase-client build - removing to rely on --no-build-isolation
google-adk==0.2.0
google-generativeai==0.5.4
# google-ai-generativelanguage==0.6.0 # Old version
google-ai-generativelanguage==0.6.4 # Updated to match google-generativeai dependency
